+++ src/mako/security/crypto/exceptions/CryptoException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\security\crypto\exceptions;
+
+use mako\security\exceptions\SecurityException;
+
+/**
+ * Crypto exception.
+ */
+class CryptoException extends SecurityException
+{
+
+}