+++ src/mako/security/password/exceptions/HasherException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\security\password\exceptions;
+
+use mako\security\exceptions\SecurityException;
+
+/**
+ * Hasher exception.
+ */
+class HasherException extends SecurityException
+{
+
+}